pyproject.toml
[tool.poetry]
name = "test-google-ads-api"
version = "0.1.0"
description = ""
authors = ["takizee <takizee53@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
google-ads = "^14.1.0"
jupyter = "^1.0.0"
pandas = "^1.3.4"
openpyxl = "^3.0.9"
pygsheets = "^2.0.5"
gspread = "^4.0.1"
gspread-dataframe = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
